Apply feedback

+++ src/plugin.validation.dns.simply/Simply/SimplyDnsClient.cs
@@ -1,5 +1,4 @@
-﻿using PKISharp.WACS.Services;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Net;
@@ -13,28 +12,23 @@ namespace PKISharp.WACS.Plugins.ValidationPlugins.Simply
 {
     public class SimplyDnsClient
     {
-        private readonly ILogService _log;
-        private readonly HttpClient _httpClient = new();
+        private readonly HttpClient _httpClient;
         private readonly string _baseUrl = "https://api.simply.com/2";
 
-        public SimplyDnsClient(string account, string apiKey, ILogService log)
+        public SimplyDnsClient(string account, string apiKey, HttpClient httpClient)
         {
-            _log = log;
+            _httpClient = httpClient;
             _httpClient.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Basic", EncodeBasicAuth(account, apiKey));
         }
 
-        public async Task CreateRecordAsync(string identifier, string domain, string value)
+        public async Task<List<Product>> GetAllProducts()
         {
-            var objectId = GetObjectId(identifier);
-            var records = await GetRecordsAsync(objectId);
+            var products = await GetProductsAsync();
+            return products;
+        }
 
-            // Remove any dangling records
-            var existingRecords = records.Where(x => x.Type == "TXT" && x.GetHostname(objectId) == domain).ToList();
-            foreach (var record in existingRecords)
-            {
-                await DeleteRecordAsync(objectId, record.RecordId);
-            }
-
+        public async Task CreateRecordAsync(string objectId, string domain, string value)
+        {
             await CreateRecordAsync(objectId, new DnsRecord
             {
                 Type = "TXT",
@@ -44,18 +38,30 @@ namespace PKISharp.WACS.Plugins.ValidationPlugins.Simply
             });
         }
 
-        public async Task DeleteRecordAsync(string identifier, string domain, string value)
+        public async Task DeleteRecordAsync(string objectId, string domain, string value)
         {
-            var objectId = GetObjectId(identifier);
+            var products = await GetProductsAsync();
+            var product = products.FirstOrDefault(x => x.Object == objectId);
+            if (product == null)
+            {
+                throw new Exception($"Unable to find product with object id {objectId}.");
+            }
             var records = await GetRecordsAsync(objectId);
-            var record = records.SingleOrDefault(x => x.Type == "TXT" && x.GetHostname(objectId) == domain && x.Data == value);
+            var record = records.SingleOrDefault(x => x.Type == "TXT" && $"{x.Name}.{product.Domain.NameIdn}" == domain && x.Data == value);
             if (record is null)
             {
-                _log.Warning($"The TXT record {domain} that should be deleted does not exist at Simply.");
-                return;
+                throw new Exception($"The TXT record {domain} that should be deleted does not exist at Simply.");
             }
 
             await DeleteRecordAsync(objectId, record.RecordId);
+        }
+
+        private async Task<List<Product>> GetProductsAsync()
+        {
+            using var response = await _httpClient.GetAsync(_baseUrl + "/my/products");
+            await using var stream = await response.Content.ReadAsStreamAsync();
+            var products = await JsonSerializer.DeserializeAsync<ProductList>(stream);
+            return products!.Products;
         }
 
         private async Task CreateRecordAsync(string objectId, DnsRecord record)
@@ -63,7 +69,6 @@ namespace PKISharp.WACS.Plugins.ValidationPlugins.Simply
             using var content = new StringContent(JsonSerializer.Serialize(record), Encoding.UTF8, "application/json");
             using var response = await _httpClient.PostAsync(_baseUrl + $"/my/products/{WebUtility.UrlEncode(objectId)}/dns/records", content);
             var responseBody = await response.Content.ReadAsStringAsync();
-            _log.Information("Simply responded with status {0}: {1}", (int)response.StatusCode, responseBody);
             response.EnsureSuccessStatusCode();
         }
 
@@ -71,30 +76,21 @@ namespace PKISharp.WACS.Plugins.ValidationPlugins.Simply
         {
             using var response = await _httpClient.DeleteAsync(_baseUrl + $"/my/products/{WebUtility.UrlEncode(objectId)}/dns/records/{recordId}");
             var responseBody = await response.Content.ReadAsStringAsync();
-            _log.Information("Simply responded with status {0}: {1}", (int)response.StatusCode, responseBody);
             response.EnsureSuccessStatusCode();
         }
 
         private async Task<List<DnsRecord>> GetRecordsAsync(string objectId)
         {
             using var response = await _httpClient.GetAsync(_baseUrl + $"/my/products/{WebUtility.UrlEncode(objectId)}/dns/records");
-            _log.Information("Simply responded with status {0}", (int)response.StatusCode);
             response.EnsureSuccessStatusCode();
             await using var stream = await response.Content.ReadAsStreamAsync();
             var records = await JsonSerializer.DeserializeAsync<DnsRecordList>(stream);
             return records!.Records;
         }
 
-
         private static string EncodeBasicAuth(string account, string apiKey)
         {
             return Convert.ToBase64String(Encoding.UTF8.GetBytes(WebUtility.UrlEncode(account) + ":" + WebUtility.UrlEncode(apiKey)));
-        }
-
-        private static string GetObjectId(string identifier)
-        {
-            var parts = identifier.Split('.');
-            return parts[^2] + "." + parts[^1];
         }
     }
 }
